# refactor: optimize performance of `queryBtcTimeLockTxByBtcTx()`