@Dawn-githup 2024-08-22
### Performance Bottleneck Optimization Verification (L1—L2)

**Before Optimization:**
- 745 ms
- 755 ms
- 635 ms
- 602 ms
- 580 ms

**After Optimization:**
- 335 ms
- 332 ms
- 349 ms
- 340 ms
- 337 ms